Golden Cookie Timer
Updated the Golden Cookie timer to show a bar at the bottom. Also
hopefully fixed the updater, unsure yet. Purposefully leaving the bars
separated so that can be another bump to demonstrate the update. Also
fixed the checking to be consistent about timing, instead of polling
every frame or so.